Don't use Diplay to convert a type to a header value
Sometimes we might want to have a header contain a value that doesn't
implement Display (HeaderName for example) which means that we couldn't
use the headers! macro.  Instead this adds a new trait that needs to be
implemented by header values. Unfortunately, since specialization is not
yet stable, we can't have a blanket impl for T: Display and only
specialize the cases we want.